Added checkmarks for all completed subtasks and tasks.

+++ tasks/db/migrate/20160701022836_create_subtasks.rb
@@ -2,7 +2,7 @@ class CreateSubtasks < ActiveRecord::Migration
   def change
     create_table :subtasks do |t|
       t.references :task, index: true, foreign_key: true
-      t.integer :percent, null: false
+      t.float :percent, null: false
       t.string :name, null: false
       t.boolean :complete, default: false
       t.timestamps null: false
